DEV: support skipping db
We need to be able to boot all the initializers
without access to a db

+++ plugin.rb
@@ -41,11 +41,15 @@ after_initialize do
   load File.expand_path("../app/controllers/discourse_cakeday/anniversaries_controller.rb", __FILE__)
   load File.expand_path("../app/controllers/discourse_cakeday/birthdays_controller.rb", __FILE__)
 
-  begin
-    if OnceoffLog.where(job_name: 'MigrateDateOfBirthToUsersTable').exists?
-      UserCustomField.where(name: 'date_of_birth').delete_all
+  skip_db = defined?(GlobalSettings.skip_db?) && GlobalSettings.skip_db?
+
+  if !skip_db
+    begin
+      if OnceoffLog.where(job_name: 'MigrateDateOfBirthToUsersTable').exists?
+        UserCustomField.where(name: 'date_of_birth').delete_all
+      end
+    rescue ActiveRecord::StatementInvalid
     end
-  rescue ActiveRecord::StatementInvalid
   end
 
   require_dependency 'user'
